requirements.txt
opencv-python
monai
nibabel
numpy
pandas
simpleitk
streamlit
streamlit-drawable-canvas
streamlit-plotly-events
streamlit_image_annotation
pydicom
pillow
scipy
scikit-image
scikit-learn